Redirect back to the originating page after editing GO code
- Now you are redirected back to the page you came from after editing.
- Set messages to be generated when codes are edited and to be displayed
when you are redirected back to pages from which editing is possible.
- Added an Edit button to info page.
- Turned the results in my_codes.php and all_codes.php into ul instead
of just a series of anchors.
- Now you can't add admins or aliases that mimic existing ones but
contain whitespace (trim whitespace).
- Small style adjustments.

+++ update_process.php
@@ -39,26 +39,78 @@ if (isset($_SESSION['AUTH'])) {
 			
 			if (isset($_POST['alias_list'])) {
 				foreach ($_POST['alias_list'] as $current_alias) {
-					$alias = new Alias($current_alias, $_POST['code'], $_POST['institution']);	
+					$current_alias = trim($current_alias);
+					$alias = new Alias($current_alias, $_POST['code'], $_POST['institution']);
+					$_SESSION['update_message'][] = "<p class='update_message'>Alias ".$current_alias." was added to '".$code->getName()."'.</p>";
 				}
 			}
 			
 			if (isset($_POST['admin_list'])) {
 				foreach ($_POST['admin_list'] as $current_admin) {
-					$code->addUser($_SESSION["AUTH"]->getId($current_admin));
+					$current_admin = trim($current_admin);
+					if ($_SESSION["AUTH"]->getId($current_admin)) {
+						// Check to see if user is already an admin
+						$select = $connection->prepare("SELECT user FROM user_to_code WHERE user = ? AND code = ?");
+  					$select->bindValue(1, $_SESSION["AUTH"]->getId($current_admin));
+  					$select->bindValue(2, $_POST['code']);
+						$select->execute();
+						$result = $select->fetchAll();
+						if (!count($result)) {
+							$code->addUser($_SESSION["AUTH"]->getId($current_admin));
+							$_SESSION['update_message'][] = "<p class='update_message'>User ".$current_admin." was added as an admin of '".$code->getName()."'.</p>";
+						}
+						else {
+							$_SESSION['update_message'][] = "<p class='update_message'>User ".$current_admin." is already an admin of '".$code->getName()."'.</p>";
+						}
+					}
 				}
 			}
 			
 			if (isset($_POST['alias_list_del'])) {
 				foreach ($_POST['alias_list_del'] as $current_alias) {
-					$alias = new Alias($current_alias, $_POST['code'], $_POST['institution']);
-					$alias->delete();	
+					$current_alias = trim($current_alias);
+					// Check to see if alias is being added. If so, don't delete.
+					if (isset($_POST['alias_list'])) {
+						$dont_delete_current_alias = 0;
+						foreach ($_POST['alias_list'] as $add_alias) {
+							if ($add_alias == $current_alias) {
+								$dont_delete_current_alias = 1;
+							}
+							if (!$dont_delete_current_alias) {
+								$alias = new Alias($current_alias, $_POST['code'], $_POST['institution']);
+								$alias->delete();	
+								$_SESSION['update_message'][] = "<p class='update_message'>Alias ".$current_alias." was removed from '".$code->getName()."'.</p>";
+							}
+						}
+					// Otherwise delete.
+					} else {
+						$alias = new Alias($current_alias, $_POST['code'], $_POST['institution']);
+						$alias->delete();
+						$_SESSION['update_message'][] = "<p class='update_message'>Alias ".$current_alias." was removed from '".$code->getName()."'.</p>";
+					}
 				}
 			}
 			
 			if (isset($_POST['admin_list_del'])) {
 				foreach ($_POST['admin_list_del'] as $current_admin) {
-					$code->delUser($_SESSION["AUTH"]->getId($current_admin));
+					$current_admin = trim($current_admin);
+					// Check to see if admin is being added. If so, don't delete.
+					if (isset($_POST['admin_list'])) {
+						$dont_delete_current_admin = 0;
+						foreach ($_POST['admin_list'] as $add_admin) {
+							if ($add_admin == $current_admin) {
+								$dont_delete_current_admin = 1;
+							}
+							if (!$dont_delete_current_admin) {
+								$code->delUser($_SESSION["AUTH"]->getId($current_admin));
+								$_SESSION['update_message'][] = "<p class='update_message'>User ".$current_admin." was removed as an admin of '".$code->getName()."'.</p>";
+							}
+						}
+					// Otherwise delete.
+					} else {
+						$code->delUser($_SESSION["AUTH"]->getId($current_admin));
+						$_SESSION['update_message'][] = "<p class='update_message'>User ".$current_admin." was removed as an admin of '".$code->getName()."'.</p>";
+					}
 				}
 			}
 			
@@ -70,3 +122,5 @@ if (isset($_SESSION['AUTH'])) {
 	} //end if (isSuperAdmin($_SESSION['AUTH']->getId()) || $is_admin) {
 
 } //end if (isset($_SESSION['AUTH'])) {
+
+header("location: " . $_POST['url']);
